Reorganize LLM utility blocks into unified llm/ directory
- Move PromptBuilderBlock and TextParserBlock from llm_utils.py to separate files in llm/ directory
- Split llm_utils.py into prompt_builder_block.py and text_parser_block.py for better organization
- Update all imports and registry exports to maintain backward compatibility
- Move test files from llm_utils/ to llm/ directory and update imports
- Consolidate all LLM-related blocks (LLMChatBlock, PromptBuilderBlock, TextParserBlock) in single module
- All 57 tests pass, maintaining full functionality

This improves code organization by grouping all LLM blocks together and provides
cleaner separation of concerns with individual files for each block type.
